+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-04T10:08:50.814Z",
+    "2026-09-06T23:56:17.054Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {